workflow-agent/requirements.txt
pocketflow
litellm
pyyaml
python-dotenv
sentence-transformers
scikit-learn
numpy

